APIO-324 Rename all tests inside apio-architect-test to *ClientTest
