+++ apio-architect-test/src/main/java/com/liferay/apio/architect/internal/test/jaxrs/filter/AcceptLanguageEmptyFilterClientTest.java
@@ -14,12 +14,15 @@
 
 package com.liferay.apio.architect.internal.test.jaxrs.filter;
 
+import static javax.ws.rs.core.HttpHeaders.ACCEPT_LANGUAGE;
+
 import static org.hamcrest.Matchers.is;
 
 import static org.junit.Assert.assertThat;
 
-import com.liferay.apio.architect.internal.jaxrs.resource.CUSTOM;
 import com.liferay.apio.architect.internal.test.base.BaseTest;
+
+import java.util.Locale;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -34,11 +37,13 @@ import org.junit.Test;
 
 /**
  * Test suite for {@link
- * com.liferay.apio.architect.internal.jaxrs.filter.HTTPMethodOverrideFilter}.
+ * com.liferay.apio.architect.internal.jaxrs.filter.AcceptLanguageEmptyFilter}.
  *
- * @author Alejandro Hernández
+ * @author Víctor Galán
+ * @author Rubén Pulido
+ * @review
  */
-public class HTTPMethodOverrideFilterTest extends BaseTest {
+public class AcceptLanguageEmptyFilterClientTest extends BaseTest {
 
 	@BeforeClass
 	public static void setUpClass() {
@@ -48,47 +53,64 @@ public class HTTPMethodOverrideFilterTest extends BaseTest {
 	}
 
 	@Test
-	public void testDefaultHTTPMethodIsNotChanged() {
+	public void testRequestWithEmptyAcceptLanguageReturnsDefaultLocale() {
 		WebTarget webTarget = createDefaultTarget();
 
 		Response response = webTarget.path(
 			"hello"
 		).request(
+		).header(
+			ACCEPT_LANGUAGE, ""
 		).get();
 
 		assertThat(response.getStatus(), is(200));
-		assertThat(response.readEntity(String.class), is("Hello World!"));
+
+		Locale expectedLocale = Locale.getDefault();
+
+		assertThat(
+			response.readEntity(String.class),
+			is(expectedLocale.toLanguageTag()));
 	}
 
 	@Test
-	public void testNotDefaultHTTPMethodIsChanged() {
+	public void testRequestWithGivenAcceptLanguageReturnsCorrespondingLocale() {
 		WebTarget webTarget = createDefaultTarget();
 
 		Response response = webTarget.path(
 			"hello"
 		).request(
-		).method(
-			"GREET"
-		);
+		).header(
+			ACCEPT_LANGUAGE, "fr-FR"
+		).get();
+
+		assertThat(response.getStatus(), is(200));
+		assertThat(response.readEntity(String.class), is("fr-FR"));
+	}
+
+	@Test
+	public void testRequestWithInvalidAcceptLanguageReturnsCorrespondingLocale() {
+		WebTarget webTarget = createDefaultTarget();
+
+		Response response = webTarget.path(
+			"hello"
+		).request(
+		).header(
+			ACCEPT_LANGUAGE, "NON_VALID_ACCEPT_LANGUAGE"
+		).get();
 
 		assertThat(response.getStatus(), is(200));
 
-		String expected = "Hello World! Method GREET.";
-
-		assertThat(response.readEntity(String.class), is(expected));
+		assertThat(response.readEntity(String.class), is("und"));
 	}
 
 	@Path("hello")
 	public static class HelloWorldResource {
 
 		@GET
-		public String hello() {
-			return "Hello World!";
-		}
-
-		@CUSTOM
 		public String hello(@Context HttpServletRequest request) {
-			return "Hello World! Method " + request.getMethod() + ".";
+			Locale locale = request.getLocale();
+
+			return locale.toLanguageTag();
 		}
 
 	}
