+++ apio-architect-test/src/main/java/com/liferay/apio/architect/internal/test/jaxrs/resource/FormResourceClientTest.java
@@ -74,7 +74,7 @@ import org.junit.Test;
  *
  * @author Alejandro Hernández
  */
-public class FormResourceTest extends BaseTest {
+public class FormResourceClientTest extends BaseTest {
 
 	@BeforeClass
 	public static void setUpClass() {
