+++ apio-architect-test/src/main/java/com/liferay/apio/architect/internal/test/jaxrs/resource/RootResourceClientTest.java
@@ -72,7 +72,7 @@ import org.junit.Test;
  *
  * @author Alejandro Hernández
  */
-public class RootResourceTest extends BaseTest {
+public class RootResourceClientTest extends BaseTest {
 
 	@BeforeClass
 	public static void setUpClass() {
